backend/requirements.txt
# AgentYard Backend — Requirements
# Python 3.12+

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database / ORM
sqlmodel==0.0.21
sqlalchemy==2.0.36
alembic==1.14.0

# SQLite for local dev (built-in), asyncpg for postgres
aiosqlite==0.20.0
asyncpg==0.30.0
greenlet==3.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client
httpx==0.27.2

# Background tasks
arq==0.26.1
redis==5.2.1

# Config
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Rate limiting
slowapi==0.1.9

# Utilities
python-multipart==0.0.12
